i#66: Windows test suite * tweak env vars to swap between 32-bit and 64-bit compiler * search for tools instead of hardcoding paths * switched to cp instead of scp due to cygwin scp not supporting colon * only supporting cygwin for now: Unix Makefiles, make, cp
i#65: more Windows test support
* 64-bit test support: fixed some build errors; disabled other tests that
  require more work, covered by i#16
* fixed test failures for common.decode*, win32.dll, client.cbr*, client.flush


SVN-Revision: 135

+++ suite/tests/security-common/TestMemProtChg.c
@@ -48,7 +48,7 @@ static int prot_codes[7] = {
 };
 
 static void 
-do_test(char *buf, int len)
+do_test(char *buf, size_t len)
 {
     int i, j;
     char *code;
